Attempt to verify bitfield and FAM

+++ t/812_shared_arena.c
@@ -33,7 +33,7 @@ TEST {
             bail_out("Failed to allocate memory for large definition string.");
         char * p = large_def;
         for (size_t i = 0; i < num_defs; ++i)
-            p += sprintf(p, "@Type%zu = { a: int, b: int, c: int, d: int };", i);
+            p += sprintf(p, "@Type%llu = { a: int, b: int, c: int, d: int };", (unsigned long long)i);
         // This call should now succeed because the arena can grow.
         infix_status status = infix_register_types(registry, large_def);
         ok(status == INFIX_SUCCESS, "Successfully registered a large number of types without allocation failure");
